+++ app/_components/measure-canvas.tsx
@@ -21,7 +21,7 @@ import { useTransformContext } from "@/_hooks/use-transform-context";
 import { Line } from "@/_lib/interfaces/line";
 
 import { KeyCode } from "@/_lib/key-code";
-import LineMenu from "./line-menu";
+import LineMenu from "@/_components/menus/line-menu";
 import { useKeyDown } from "@/_hooks/use-key-down";
 import { useKeyUp } from "@/_hooks/use-key-up";
 import { CM } from "@/_lib/unit";
@@ -212,11 +212,11 @@ export default function MeasureCanvas({
 
   useKeyDown(() => {
     setAxisConstrained(true);
-  }, [KeyCode.ShiftLeft, KeyCode.ShiftRight]);
+  }, [KeyCode.Shift]);
 
   useKeyUp(() => {
     setAxisConstrained(false);
-  }, [KeyCode.ShiftLeft, KeyCode.ShiftRight]);
+  }, [KeyCode.Shift]);
 
   useEffect(() => {
     const canvas = canvasRef.current;